-rename all logic handlers as serivce -remove ui threads for pyqt -adjust commands_auotomator html to include llm UI elements
