+++ command_automator.pyw
@@ -6,20 +6,29 @@ import os
 import threading
 import aiofiles
 import webview
-from logic_handler import LogicHandler
+
+from services.commands_automator_service import CommandsAutomatorService
+from services.configuration_service import ConfigurationService
+from tabs.llm.llm_logic_handler import LLMService
+
 
 # logging.basicConfig(level=logging.DEBUG)
 
 class CommandsAutomatorApi:
-    def __init__(self):
-        self.logic_handler = LogicHandler()
-        self.config_path = "config/commands-executor-config.json"
+
+    lLLMLogicHanlder: LLMService = LLMService()
+
+
     
+    def __init__(self):
+        self.commands_automator_service = CommandsAutomatorService()
+        self.commands_automator_config = ConfigurationService("config/commands-executor-config.json")
+        self.llm_config = ConfigurationService('llm-config.json')
         self.loop = None
         self.executor = ThreadPoolExecutor(max_workers=4)
-        self._setup_event_loop()
+        self.setup_event_loop()
     
-    def _setup_event_loop(self):
+    def setup_event_loop(self):
         def run_loop():
             self.loop = asyncio.new_event_loop()
             asyncio.set_event_loop(self.loop)
@@ -32,17 +41,17 @@ class CommandsAutomatorApi:
             threading.Event().wait(0.01)
 
     def load_scripts(self):
-        return self.logic_handler.load_scripts()
+        return self.commands_automator_service.load_scripts()
 
     def get_script_description(self, script_name):
-        script_file = self.logic_handler.get_name_to_scripts()[script_name]
-        return self.logic_handler.get_script_description(script_file)
+        script_file = self.commands_automator_service.get_name_to_scripts()[script_name]
+        return self.commands_automator_service.get_script_description(script_file)
     
-    def load_configuration(self):
-        return self.run_async_method(self.load_configuration_async)
+    def load_commands_configuration(self):
+        return self.run_async_method(self.commands_automator_config.load_configuration_async)
 
-    def save_configuration(self, config):
-        return self.run_async_method(self.save_configuration_async, config)
+    def save_commands_configuration(self, config):
+        return self.run_async_method(self.commands_automator_config.save_configuration_async, config)
 
     def run_async_method(self, async_method, *args, **kwargs):
         try:
@@ -59,21 +68,24 @@ class CommandsAutomatorApi:
         except Exception as e:
             logging.error(f"Error running async method: {e}")
             return {"error": str(e)}
-        
-    async def load_configuration_async(self):
-        async with aiofiles.open(self.config_path, "r") as f:
-            data = await f.read()
-        return json.loads(data)
-
-    async def save_configuration_async(self, config):
-        async with aiofiles.open(self.config_path, "w") as f:
-            await f.write(json.dumps(config, indent=4))
-        
-        return True
+   
 
     def execute_script(self, script_name, additional_text, flags):
-        return self.logic_handler.execute_script(script_name, additional_text, flags)
+        return self.commands_automator_service.execute_script(script_name, additional_text, flags)
     
+
+
+    def call_llm(self, query):
+    # Implement your LLM logic here, or call your existing handler
+    # For demonstration, just echo the query
+        return "trial"
+    
+    def load_llm_configuration(self):
+        return self.run_async_method(self.llm_config.load_configuration_async)
+
+    def save_llm_configuration(self, config):
+        return self.run_async_method(self.llm_config.save_configuration_async, config)
+
 
 if __name__ == '__main__':
     api = CommandsAutomatorApi()
